+++ CHANGELOG.md
@@ -4,11 +4,12 @@
 
 ### Added
 - Sites now expose ports 3000-3005 for miscellaneous use.
-- Added the `keys` command to copy SSH keys from the host machine into a sites container.
+- Added the `keys` command to copy SSH keys from the host machine into a sites' container.
 - The default shell is now bash and not sh, allowing common scripts.
 - Site containers now include common image processing libraries.
 - Added common Linux build tools such as make to the site containers.
 - Yarn is now installed by default.
+- Added the ability to exclude directories from a site bind mounts to improve performance.
 
 ### Changed
 - The `composer` command now accepts site arguments. ([#364](https://github.com/craftcms/nitro/issues/364))